dom/media/test/test_seek-2.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Media test: seek tests</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script type="text/javascript" src="manifest.js"></script>
  <script type="text/javascript" src="seek_support.js"></script>
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

// The data being used in these tests is specified in manifest.js.
// The functions to build the test array and to run a specific test are in
//      seek_support.js.

PARALLEL_TESTS = 1;
const SEEK_TEST_NUMBER = 2;

function test_seek2(v, seekTime, is, ok, finish) {

// Test seeking works if current time is set before video is
// playing.
var startPassed = false;
var endPassed = false;
var completed = false;

function startTest() {
  if (completed)
    return;

  v.currentTime=seekTime;
  v.play();
}

function seekStarted() {
  if (completed)
    return;

  ok(v.currentTime >= seekTime - 0.1, "Video currentTime should be around " + seekTime + ": " + v.currentTime);
  startPassed = true;
}

function seekEnded() {
  if (completed)
    return;

  endPassed = true;
}

function playbackEnded() {
  if (completed)
    return;

  completed = true;
  ok(startPassed, "send seeking event");
  ok(endPassed, "send seeked event");
  ok(v.ended, "Checking playback has ended");
  ok(Math.abs(v.currentTime - v.duration) <= 0.1, "Checking currentTime at end: " + v.currentTime);
  finish();
}

v.addEventListener("ended", playbackEnded);
v.addEventListener("loadedmetadata", startTest);
v.addEventListener("seeking", seekStarted);
v.addEventListener("seeked", seekEnded);

}

manager.runTests(createTestArray(), startTest);

</script>
</pre>
</body>
</html>
